+++ jest.config.js
@@ -3,5 +3,6 @@ module.exports = {
   testEnvironment: 'node',
   testMatch: ['<rootDir>/src/test/**/*.ts'],
   testPathIgnorePatterns: ['<rootDir>/src/test/helpers'],
-  coveragePathIgnorePatterns: ['<rootDir>/src/test/helpers']
+  coveragePathIgnorePatterns: ['<rootDir>/src/test/helpers'],
+  maxWorkers: 1
 };